+++ opentelemetry-zipkin/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## vNext
 
+- **Breaking** The `ZipkinExporterBuilder::build()` method now returns a
+  `Result<ZipkinExporter, ExporterBuildError>`. The `ExporterBuildError` enum
+  lists possible failures specific to the Zipkin exporter and was renamed from
+  `opentelemetry_zipkin::Error`. Previously, this method returned a `TraceError`
+  from the `opentelemetry_sdk` crate, which was unrelated to Zipkin builder
+  failures.
+  [2839](https://github.com/open-telemetry/opentelemetry-rust/pull/2839)
+
 ## 0.28.0
 
 Released 2025-Feb-10